chore: Code split network utils class
